backend/requirements.txt
fastapi
uvicorn
pymongo
passlib[bcrypt]
bcrypt==4.0.1
python-jose
python-multipart
pydantic
email-validator
requests
torch
torchvision
pillow
timm
